Changed ExMODEL parameters to ELLIPx_MODEL. Changed EPSxMODEL parameters to POLARx_MODEL. Added AMODEL, BMODEL, THETAMODEL for both _IMAGE and _WORLD coordinates. Added CXXMODEL CYYMODEL, CXYMODEL for both _IMAGE and _WORLD coordinates. AddedTHETAMODEL_SKY,THETAMODEL_J2000,THETAMODEL_B1950. Added ELLIPxMODEL_WORLD and POLARxMODEL_WORLD. Fixed broken PSF-fitting. Fixed warnings while generating the configure (thanks to S.Pascual). Pushed version number to 2.9.1.

+++ src/filter.h
@@ -10,7 +10,7 @@
 *	Contents:	functions dealing with on-line filtering of the image
 *			(for detection).
 *
-*	Last modify:	09/11/99
+*	Last modify:	13/09/2009
 *
 *%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 */
@@ -36,8 +36,8 @@ filterstruct	*thefilter;
 
 /*------------------------------- functions ---------------------------------*/
 void		convolve(picstruct *, PIXTYPE *),
-		convolve_image(picstruct *field, double *vig1,
-				double *vig2, int width, int height),
+		convolve_image(picstruct *field, float *vig1,
+				float *vig2, int width, int height),
 		filter(picstruct *, PIXTYPE *),
 		neurfilter(picstruct *, PIXTYPE *),
 		endfilter(void),
